a little bit of structural sanity now that there are no other ppc prs out there to cause conflicts with
