+++ assets/src/edit-story/components/inspector/prepublish/components/autoVideoOptimization.js
@@ -24,15 +24,15 @@ import { trackClick } from '@web-stories-wp/tracking';
 /**
  * Internal dependencies
  */
-import { Link, THEME_CONSTANTS } from '../../../../design-system';
-import { useConfig } from '../../../app';
+import { Link, THEME_CONSTANTS } from '../../../../../design-system';
+import { useConfig } from '../../../../app';
 import {
   DescriptionText,
   PageIndicator,
   StyledToggle,
   ToggleGroup,
   VideoOptimizationGroup,
-} from './styles';
+} from '../styles';
 
 const AutoVideoOptimization = ({
   areVideosAutoOptimized,
