+++ assets/src/edit-story/components/inspector/prepublish/components/reviewChecklistDialog.js
@@ -23,8 +23,8 @@ import { __ } from '@web-stories-wp/i18n';
 /**
  * Internal dependencies
  */
-import { Text, THEME_CONSTANTS } from '../../../../design-system';
-import Dialog from '../../dialog';
+import { Text, THEME_CONSTANTS } from '../../../../../design-system';
+import Dialog from '../../../dialog';
 
 const TEXT = {
   TITLE: __('Review checklist before publishing.', 'web-stories'),
